Add draw.js and refactor main.js to use it
Created a new class 'Draw' in a separate file 'draw.js' that encapsulates the drawing logic and state of canvas. 'main.js' was refactored to import and use the 'Draw' class instead of directly manipulating the canvas. This change improves code readability and maintainability. The favicon link in index.html is commented out, and script linking classPractice.js is removed.