Add bouncing balls feature and clean up Draw methods
This commit introduces the bouncing balls animation in a separate JavaScript file and a corresponding HTML page for the same. The Draw class has been updated to improve the circle and clear methods for more flexibility and better readability. Also, removed the unnecessary code in main.js and index.html to keep the files clear and understandable. A clear() method has been added to the Draw class to easily clear the canvas. The circle method in the Draw class has been updated to fill the circle with color. In addition, adjustments have been made in the config file for proper routing of the new page.