Add background color & refactor circle drawing method
The commit adds a property `bgColor` to the Draw class that remembers the last color set for the background. This enables the `clear` method to maintain the defined background color when clearing the canvas. In the circle drawing method, commenting off the fill function is to maintain the color of the background when drawing the circle. In `main.js` and `BouncingBall.js`, code was updated to include random color creation for each circle and motion, improving the dynamics and colorfulness of the animation.